S02P23B204-56, S02P23B204-59|Update Entity Columns Modified

+++ back/findme/src/main/java/com/ssafy/findme/repository/SympRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ssafy.findme.domain.Symp;
 
-public interface SympRepository extends JpaRepository<Symp, Integer>{
+public interface SympRepository extends JpaRepository<Symp, Long>{
 	List<Symp> findByUser_Id(Long user_id);
 	Symp findByReview_IdAndUser_Id(Long review_id, Long user_id);
 }
